+++ src/aks-preview/azext_aks_preview/decorator.py
@@ -65,7 +65,10 @@ from azext_aks_preview.addonconfiguration import (
     ensure_container_insights_for_monitoring,
     ensure_default_log_analytics_workspace_for_monitoring,
 )
-from azext_aks_preview.custom import _get_snapshot
+from azext_aks_preview.custom import (
+    _get_snapshot,
+    _get_cluster_snapshot,
+)
 
 logger = get_logger(__name__)
 
@@ -73,7 +76,8 @@ logger = get_logger(__name__)
 ContainerServiceClient = TypeVar("ContainerServiceClient")
 Identity = TypeVar("Identity")
 ManagedCluster = TypeVar("ManagedCluster")
-ManagedClusterLoadBalancerProfile = TypeVar("ManagedClusterLoadBalancerProfile")
+ManagedClusterLoadBalancerProfile = TypeVar(
+    "ManagedClusterLoadBalancerProfile")
 ResourceReference = TypeVar("ResourceReference")
 KubeletConfig = TypeVar("KubeletConfig")
 LinuxOSConfig = TypeVar("LinuxOSConfig")
@@ -83,6 +87,7 @@ ManagedClusterAddonProfile = TypeVar("ManagedClusterAddonProfile")
 ManagedClusterOIDCIssuerProfile = TypeVar('ManagedClusterOIDCIssuerProfile')
 ManagedClusterSecurityProfileWorkloadIdentity = TypeVar('ManagedClusterSecurityProfileWorkloadIdentity')
 Snapshot = TypeVar("Snapshot")
+ManagedClusterSnapshot = TypeVar("ManagedClusterSnapshot")
 AzureKeyVaultKms = TypeVar('AzureKeyVaultKms')
 
 
@@ -356,7 +361,8 @@ class AKSPreviewContext(AKSContext):
             )
             if (
                 agent_pool_profile and
-                hasattr(agent_pool_profile, "workload_runtime") and  # backward compatibility
+                # backward compatibility
+                hasattr(agent_pool_profile, "workload_runtime") and
                 agent_pool_profile.workload_runtime is not None
             ):
                 workload_runtime = agent_pool_profile.workload_runtime
@@ -379,7 +385,8 @@ class AKSPreviewContext(AKSContext):
             )
             if (
                 agent_pool_profile and
-                hasattr(agent_pool_profile, "gpu_instance_profile") and  # backward compatibility
+                # backward compatibility
+                hasattr(agent_pool_profile, "gpu_instance_profile") and
                 agent_pool_profile.gpu_instance_profile is not None
             ):
                 gpu_instance_profile = agent_pool_profile.gpu_instance_profile
@@ -404,8 +411,10 @@ class AKSPreviewContext(AKSContext):
                         message_of_the_day_file_path
                     )
                 )
-            message_of_the_day = read_file_content(message_of_the_day_file_path)
-            message_of_the_day = base64.b64encode(bytes(message_of_the_day, 'ascii')).decode('ascii')
+            message_of_the_day = read_file_content(
+                message_of_the_day_file_path)
+            message_of_the_day = base64.b64encode(
+                bytes(message_of_the_day, 'ascii')).decode('ascii')
 
         # try to read the property value corresponding to the parameter from the `mc` object
         if self.mc and self.mc.agent_pool_profiles:
@@ -414,7 +423,8 @@ class AKSPreviewContext(AKSContext):
             )
             if (
                 agent_pool_profile and
-                hasattr(agent_pool_profile, "message_of_the_day") and  # backward compatibility
+                # backward compatibility
+                hasattr(agent_pool_profile, "message_of_the_day") and
                 agent_pool_profile.message_of_the_day is not None
             ):
                 message_of_the_day = agent_pool_profile.message_of_the_day
@@ -560,7 +570,8 @@ class AKSPreviewContext(AKSContext):
         :return: int or None
         """
         # read the original value passed by the command
-        nat_gateway_managed_outbound_ip_count = self.raw_param.get("nat_gateway_managed_outbound_ip_count")
+        nat_gateway_managed_outbound_ip_count = self.raw_param.get(
+            "nat_gateway_managed_outbound_ip_count")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -586,7 +597,8 @@ class AKSPreviewContext(AKSContext):
         :return: int or None
         """
         # read the original value passed by the command
-        nat_gateway_idle_timeout = self.raw_param.get("nat_gateway_idle_timeout")
+        nat_gateway_idle_timeout = self.raw_param.get(
+            "nat_gateway_idle_timeout")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -612,7 +624,8 @@ class AKSPreviewContext(AKSContext):
         :return: bool
         """
         # read the original value passed by the command
-        enable_pod_security_policy = self.raw_param.get("enable_pod_security_policy")
+        enable_pod_security_policy = self.raw_param.get(
+            "enable_pod_security_policy")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -650,7 +663,8 @@ class AKSPreviewContext(AKSContext):
         :return: bool
         """
         # read the original value passed by the command
-        disable_pod_security_policy = self.raw_param.get("disable_pod_security_policy")
+        disable_pod_security_policy = self.raw_param.get(
+            "disable_pod_security_policy")
         # We do not support this option in create mode, therefore we do not read the value from `mc`.
 
         # this parameter does not need dynamic completion
@@ -686,7 +700,8 @@ class AKSPreviewContext(AKSContext):
 
         :return: bool
         """
-        enable_managed_identity = super()._get_enable_managed_identity(enable_validation, read_only)
+        enable_managed_identity = super()._get_enable_managed_identity(
+            enable_validation, read_only)
         # additional validation
         if enable_validation:
             if self.decorator_mode == DecoratorMode.CREATE:
@@ -807,7 +822,8 @@ class AKSPreviewContext(AKSContext):
         :return: bool
         """
         # read the original value passed by the command
-        enable_pod_identity_with_kubenet = self.raw_param.get("enable_pod_identity_with_kubenet")
+        enable_pod_identity_with_kubenet = self.raw_param.get(
+            "enable_pod_identity_with_kubenet")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -868,8 +884,10 @@ class AKSPreviewContext(AKSContext):
         """
         # determine the value of constants
         addon_consts = self.get_addon_consts()
-        CONST_INGRESS_APPGW_ADDON_NAME = addon_consts.get("CONST_INGRESS_APPGW_ADDON_NAME")
-        CONST_INGRESS_APPGW_SUBNET_CIDR = addon_consts.get("CONST_INGRESS_APPGW_SUBNET_CIDR")
+        CONST_INGRESS_APPGW_ADDON_NAME = addon_consts.get(
+            "CONST_INGRESS_APPGW_ADDON_NAME")
+        CONST_INGRESS_APPGW_SUBNET_CIDR = addon_consts.get(
+            "CONST_INGRESS_APPGW_SUBNET_CIDR")
 
         # read the original value passed by the command
         appgw_subnet_prefix = self.raw_param.get("appgw_subnet_prefix")
@@ -900,11 +918,14 @@ class AKSPreviewContext(AKSContext):
         """
         # determine the value of constants
         addon_consts = self.get_addon_consts()
-        CONST_MONITORING_ADDON_NAME = addon_consts.get("CONST_MONITORING_ADDON_NAME")
-        CONST_MONITORING_USING_AAD_MSI_AUTH = addon_consts.get("CONST_MONITORING_USING_AAD_MSI_AUTH")
+        CONST_MONITORING_ADDON_NAME = addon_consts.get(
+            "CONST_MONITORING_ADDON_NAME")
+        CONST_MONITORING_USING_AAD_MSI_AUTH = addon_consts.get(
+            "CONST_MONITORING_USING_AAD_MSI_AUTH")
 
         # read the original value passed by the command
-        enable_msi_auth_for_monitoring = self.raw_param.get("enable_msi_auth_for_monitoring")
+        enable_msi_auth_for_monitoring = self.raw_param.get(
+            "enable_msi_auth_for_monitoring")
         # try to read the property value corresponding to the parameter from the `mc` object
         if (
             self.mc and
@@ -935,7 +956,8 @@ class AKSPreviewContext(AKSContext):
         no_wait = super().get_no_wait()
 
         if self.get_intermediate("monitoring") and self.get_enable_msi_auth_for_monitoring():
-            logger.warning("Enabling msi auth for monitoring addon requires waiting for cluster creation to complete")
+            logger.warning(
+                "Enabling msi auth for monitoring addon requires waiting for cluster creation to complete")
             if no_wait:
                 logger.warning("The set option '--no-wait' has been ignored")
                 no_wait = False
@@ -962,7 +984,8 @@ class AKSPreviewContext(AKSContext):
         """
         # determine the value of constants
         addon_consts = self.get_addon_consts()
-        CONST_MONITORING_ADDON_NAME = addon_consts.get("CONST_MONITORING_ADDON_NAME")
+        CONST_MONITORING_ADDON_NAME = addon_consts.get(
+            "CONST_MONITORING_ADDON_NAME")
         CONST_MONITORING_LOG_ANALYTICS_WORKSPACE_RESOURCE_ID = addon_consts.get(
             "CONST_MONITORING_LOG_ANALYTICS_WORKSPACE_RESOURCE_ID"
         )
@@ -1201,7 +1224,8 @@ class AKSPreviewContext(AKSContext):
                 # not been decorated into the mc object at this time, only the value after dynamic completion is
                 # meaningful here.
                 if safe_lower(self._get_load_balancer_sku(enable_validation=False)) == "basic":
-                    raise InvalidArgumentValueError("{} doesn't support basic load balancer sku".format(outbound_type))
+                    raise InvalidArgumentValueError(
+                        "{} doesn't support basic load balancer sku".format(outbound_type))
                 if outbound_type == CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY:
                     if self.get_vnet_subnet_id() in ["", None]:
                         raise RequiredArgumentMissingError(
@@ -1251,7 +1275,8 @@ class AKSPreviewContext(AKSContext):
             if (
                 self.mc and
                 self.mc.windows_profile and
-                hasattr(self.mc.windows_profile, "gmsa_profile") and  # backward compatibility
+                # backward compatibility
+                hasattr(self.mc.windows_profile, "gmsa_profile") and
                 self.mc.windows_profile.gmsa_profile and
                 self.mc.windows_profile.gmsa_profile.enabled is not None
             ):
@@ -1302,7 +1327,8 @@ class AKSPreviewContext(AKSContext):
             if (
                 self.mc and
                 self.mc.windows_profile and
-                hasattr(self.mc.windows_profile, "gmsa_profile") and  # backward compatibility
+                # backward compatibility
+                hasattr(self.mc.windows_profile, "gmsa_profile") and
                 self.mc.windows_profile.gmsa_profile and
                 self.mc.windows_profile.gmsa_profile.dns_server is not None
             ):
@@ -1318,7 +1344,8 @@ class AKSPreviewContext(AKSContext):
             if (
                 self.mc and
                 self.mc.windows_profile and
-                hasattr(self.mc.windows_profile, "gmsa_profile") and  # backward compatibility
+                # backward compatibility
+                hasattr(self.mc.windows_profile, "gmsa_profile") and
                 self.mc.windows_profile.gmsa_profile and
                 self.mc.windows_profile.gmsa_profile.root_domain_name is not None
             ):
@@ -1404,6 +1431,50 @@ class AKSPreviewContext(AKSContext):
             self.set_intermediate("snapshot", snapshot, overwrite_exists=True)
         return snapshot
 
+    def get_cluster_snapshot_id(self) -> Union[str, None]:
+        """Obtain the values of cluster_snapshot_id.
+
+        :return: string or None
+        """
+        # read the original value passed by the command
+        snapshot_id = self.raw_param.get("cluster_snapshot_id")
+        # try to read the property value corresponding to the parameter from the `mc` object
+        if (
+            self.mc and
+            self.mc.creation_data and
+            self.mc.creation_data.source_resource_id is not None
+        ):
+            snapshot_id = (
+                self.mc.creation_data.source_resource_id
+            )
+
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        return snapshot_id
+
+    def get_cluster_snapshot(self) -> Union[ManagedClusterSnapshot, None]:
+        """Helper function to retrieve the ManagedClusterSnapshot object corresponding to a cluster snapshot id.
+
+        This fuction will store an intermediate "managedclustersnapshot" to avoid sending the same request multiple times.
+
+        Function "_get_cluster_snapshot" will be called to retrieve the ManagedClusterSnapshot object corresponding to a cluster snapshot id, which
+        internally used the managedclustersnapshot client (managedclustersnapshots operations belonging to container service client) to send
+        the request.
+
+        :return: ManagedClusterSnapshot or None
+        """
+        # try to read from intermediates
+        snapshot = self.get_intermediate("managedclustersnapshot")
+        if snapshot:
+            return snapshot
+
+        snapshot_id = self.get_cluster_snapshot_id()
+        if snapshot_id:
+            snapshot = _get_cluster_snapshot(self.cmd.cli_ctx, snapshot_id)
+            self.set_intermediate("managedclustersnapshot",
+                                  snapshot, overwrite_exists=True)
+        return snapshot
+
     def get_host_group_id(self) -> Union[str, None]:
         return self._get_host_group_id()
 
@@ -1439,11 +1510,17 @@ class AKSPreviewContext(AKSContext):
             value_obtained_from_mc = self.mc.kubernetes_version
         # try to retrieve the value from snapshot
         value_obtained_from_snapshot = None
+        value_obtained_from_cluster_snapshot = None
         # skip dynamic completion if read_only is specified
         if not read_only:
             snapshot = self.get_snapshot()
             if snapshot:
                 value_obtained_from_snapshot = snapshot.kubernetes_version
+
+        if not read_only:
+            snapshot = self.get_cluster_snapshot()
+            if snapshot:
+                value_obtained_from_cluster_snapshot = snapshot.managed_cluster_properties_read_only.kubernetes_version
 
         # set default value
         if value_obtained_from_mc is not None:
@@ -1451,6 +1528,8 @@ class AKSPreviewContext(AKSContext):
         # default value is an empty string
         elif raw_value:
             kubernetes_version = raw_value
+        elif not read_only and value_obtained_from_cluster_snapshot is not None:
+            kubernetes_version = value_obtained_from_cluster_snapshot
         elif not read_only and value_obtained_from_snapshot is not None:
             kubernetes_version = value_obtained_from_snapshot
         else:
@@ -1654,7 +1733,8 @@ class AKSPreviewContext(AKSContext):
         """
         # read the original value passed by the command
         # TODO: set default value as False after the get function of AKSParamDict accepts parameter `default`
-        enable_azure_keyvault_kms = self.raw_param.get("enable_azure_keyvault_kms")
+        enable_azure_keyvault_kms = self.raw_param.get(
+            "enable_azure_keyvault_kms")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -1693,7 +1773,8 @@ class AKSPreviewContext(AKSContext):
         :return: string or None
         """
         # read the original value passed by the command
-        azure_keyvault_kms_key_id = self.raw_param.get("azure_keyvault_kms_key_id")
+        azure_keyvault_kms_key_id = self.raw_param.get(
+            "azure_keyvault_kms_key_id")
         # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
         if self.decorator_mode == DecoratorMode.CREATE:
             if (
@@ -1705,7 +1786,8 @@ class AKSPreviewContext(AKSContext):
                 azure_keyvault_kms_key_id = self.mc.security_profile.azure_key_vault_kms.key_id
 
         if enable_validation:
-            enable_azure_keyvault_kms = self._get_enable_azure_keyvault_kms(enable_validation=False)
+            enable_azure_keyvault_kms = self._get_enable_azure_keyvault_kms(
+                enable_validation=False)
             if (
                 azure_keyvault_kms_key_id and
                 (
@@ -1713,7 +1795,8 @@ class AKSPreviewContext(AKSContext):
                     enable_azure_keyvault_kms is False
                 )
             ):
-                raise RequiredArgumentMissingError('"--azure-keyvault-kms-key-id" requires "--enable-azure-keyvault-kms".')
+                raise RequiredArgumentMissingError(
+                    '"--azure-keyvault-kms-key-id" requires "--enable-azure-keyvault-kms".')
 
         return azure_keyvault_kms_key_id
 
@@ -1793,6 +1876,23 @@ class AKSPreviewCreateDecorator(AKSCreateDecorator):
         agent_pool_profile.capacity_reservation_group_id = self.context.get_crg_id()
 
         mc.agent_pool_profiles = [agent_pool_profile]
+        return mc
+
+    def set_up_creationdata_of_cluster_snapshot(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up creationData of cluster snapshot for the ManagedCluster object.
+
+        Note: Inherited and extended in aks-preview to set some additional properties.
+
+        :return: the ManagedCluster object
+        """
+        # snapshot creation data
+        creation_data = None
+        snapshot_id = self.context.get_cluster_snapshot_id()
+        if snapshot_id:
+            creation_data = self.models.CreationData(
+                source_resource_id=snapshot_id
+            )
+        mc.creation_data = creation_data
         return mc
 
     def set_up_http_proxy_config(self, mc: ManagedCluster) -> ManagedCluster:
@@ -1961,7 +2061,8 @@ class AKSPreviewCreateDecorator(AKSCreateDecorator):
             create_dcra=False,
         )
         # set intermediate
-        self.context.set_intermediate("monitoring", True, overwrite_exists=True)
+        self.context.set_intermediate(
+            "monitoring", True, overwrite_exists=True)
         return monitoring_addon_profile
 
     def build_ingress_appgw_addon_profile(self) -> ManagedClusterAddonProfile:
@@ -2111,6 +2212,7 @@ class AKSPreviewCreateDecorator(AKSCreateDecorator):
         mc = self.set_up_oidc_issuer_profile(mc)
 
         mc = self.set_up_azure_keyvault_kms(mc)
+        mc = self.set_up_creationdata_of_cluster_snapshot(mc)
         return mc
 
     def create_mc_preview(self, mc: ManagedCluster) -> ManagedCluster:
@@ -2125,7 +2227,8 @@ class AKSPreviewCreateDecorator(AKSCreateDecorator):
 
         # determine the value of constants
         addon_consts = self.context.get_addon_consts()
-        CONST_MONITORING_ADDON_NAME = addon_consts.get("CONST_MONITORING_ADDON_NAME")
+        CONST_MONITORING_ADDON_NAME = addon_consts.get(
+            "CONST_MONITORING_ADDON_NAME")
 
         # Due to SPN replication latency, we do a few retries here
         max_retry = 30
@@ -2201,7 +2304,8 @@ class AKSPreviewUpdateDecorator(AKSUpdateDecorator):
         excluded_keys = ("cmd", "client", "resource_group_name", "name")
         # check whether the remaining parameters are set
         # the default value None or False (and other empty values, like empty string) will be considered as not set
-        is_changed = any(v for k, v in self.context.raw_param.items() if k not in excluded_keys)
+        is_changed = any(
+            v for k, v in self.context.raw_param.items() if k not in excluded_keys)
 
         # special cases
         # some parameters support the use of empty string or dictionary to update/remove previously set values
@@ -2387,7 +2491,8 @@ class AKSPreviewUpdateDecorator(AKSUpdateDecorator):
                 )
 
         if self.context.get_disable_pod_identity():
-            _update_addon_pod_identity(mc, enable=False, models=self.models.pod_identity_models)
+            _update_addon_pod_identity(
+                mc, enable=False, models=self.models.pod_identity_models)
         return mc
 
     def update_oidc_issuer_profile(self, mc: ManagedCluster) -> ManagedCluster: